couldn't be happier with this :DDD

+++ rumors/src/main/java/org/columbia/controller/RumorController.java
@@ -1,9 +1,11 @@
 package org.columbia.controller;
 
 import org.columbia.dto.RumorDto;
+import org.columbia.dto.RumorIdDto;
 import org.columbia.entity.RumorEntity;
 import org.columbia.service.RumorService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
@@ -26,6 +28,11 @@ public class RumorController {
     @RequestMapping(value = "/rumor/{id}", method = RequestMethod.GET)
     @ResponseBody
     public ResponseEntity<?> getRumorById(@PathVariable UUID id) {
+
+        if (!rumorService.rumorExistsBbyId(id)) {
+            return new ResponseEntity<>("Not Found", HttpStatus.NOT_FOUND);
+        }
+
         RumorEntity rumor = rumorService.getRumorByID(id);
         return ResponseEntity.ok(rumorService.convertToRumorIdDto(rumor));
     }
@@ -35,12 +42,17 @@ public class RumorController {
     public ResponseEntity<?> postRumor(@RequestBody RumorDto rumor) {
         RumorEntity entity = rumorService.convertToRumorEntity(rumor);
         rumorService.createRumor(entity);
-        return ResponseEntity.ok(rumorService.convertToRumorIdDto(entity));
+        return new ResponseEntity<RumorIdDto>(rumorService.convertToRumorIdDto(entity), HttpStatus.CREATED);
     }
 
     @RequestMapping(value = "/rumor/{id}", method = RequestMethod.DELETE)
     @ResponseBody
     public ResponseEntity<?> deleteRumor(@PathVariable UUID id) {
+
+        if (!rumorService.rumorExistsBbyId(id)) {
+            return new ResponseEntity<>("Not Found", HttpStatus.NOT_FOUND);
+        }
+
         rumorService.deleteRumorbyId(id);
         return ResponseEntity.noContent().build();
     }
